Refactor password handling for serverless deployment
